xdire/base-lb-forwarder-impl fix: - conn close in many instances was not created - zerolog logging - error handling - naming - tests in the wrong place - context optimizations - concurrency locks
