+++ forwarder.go
@@ -2,13 +2,20 @@ package xlb
 
 import (
 	"context"
-	"crypto/tls"
+	"errors"
 	"fmt"
+	"github.com/rs/zerolog"
 	"io"
+	"math"
 	"net"
+	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
+)
+
+const (
+	closedNetworkConnection = "use of closed network connection"
 )
 
 type route struct {
@@ -23,11 +30,15 @@ type Forwarder struct {
 	mutex       sync.RWMutex
 	updateLock  bool
 	strategy    leastConnection
-	logger      Logger
+	logger      zerolog.Logger
 	dialTimeout time.Duration
 }
 
-func NewForwarder(ctx context.Context, params ServicePool, logger Logger) *Forwarder {
+// NewForwarder creates load balancer forwarder that can be used to
+// establish proxy communication between client and destination, adding
+// some features like: the least loaded server balancing, health check
+// routing, dynamic route update
+func NewForwarder(params ServicePool, logger zerolog.Logger) *Forwarder {
 	fwd := &Forwarder{
 		routes: &[]*route{},
 		logger: logger,
@@ -57,6 +68,8 @@ func NewForwarder(ctx context.Context, params ServicePool, logger Logger) *Forwa
 	return fwd
 }
 
+// UpdateServicePool will update service pool merging the new pool routes
+// configuration with existing routes
 func (f *Forwarder) UpdateServicePool(pool ServicePool) {
 	f.mutex.Lock()
 	defer f.mutex.Unlock()
@@ -94,18 +107,14 @@ func (f *Forwarder) UpdateServicePool(pool ServicePool) {
 		rte.active.Store(false)
 	}
 	f.routes = &newRoutePool
-	f.logger.Info(fmt.Sprintf("forwarder routes updated to: %+v from: %+v", *f.routes, pool.Routes()))
+	f.logger.Info().Msgf("forwarder routes updated to: %+v from: %+v", *f.routes, pool.Routes())
 }
 
-func (f *Forwarder) Attach(ctx context.Context, in *tls.Conn) error {
+// Attach will attach some incoming session to the pool of upstream traffic distribution
+func (f *Forwarder) Attach(ctx context.Context, in io.ReadWriteCloser) error {
 
 	errTransport := make(chan error, 2)
-	defer func(conn net.Conn) {
-		err := conn.Close()
-		if err != nil {
-			f.logger.Debug(fmt.Sprintf("forwarder conn closed: %s", in.LocalAddr().String()))
-		}
-	}(in)
+	defer in.Close()
 
 	rte := f.strategy.Next()
 	if rte == nil {
@@ -114,19 +123,24 @@ func (f *Forwarder) Attach(ctx context.Context, in *tls.Conn) error {
 
 	dest, err := net.DialTimeout("tcp", rte.address, f.dialTimeout)
 	if err != nil {
-		f.logger.Error(fmt.Sprintf("route unreachable %s", rte.address))
+		f.logger.Err(err).Msgf("route unreachable %s", rte.address)
 		return err
 	}
+	defer dest.Close()
 
 	// Connection increment here as we reached destination
 	atomic.AddUint32(&rte.connections, 1)
 
-	go func(w io.Writer, r io.Reader) {
+	go func(w io.WriteCloser, r io.ReadCloser) {
+		defer w.Close()
+		defer r.Close()
 		_, err := io.Copy(w, r)
 		errTransport <- err
 	}(dest, in)
 
-	go func(w io.Writer, r io.Reader) {
+	go func(w io.WriteCloser, r io.ReadCloser) {
+		defer w.Close()
+		defer r.Close()
 		_, err := io.Copy(w, r)
 		errTransport <- err
 	}(in, dest)
@@ -137,7 +151,9 @@ func (f *Forwarder) Attach(ctx context.Context, in *tls.Conn) error {
 		case <-ctx.Done():
 			return ctx.Err()
 		case err = <-errTransport:
-			if err != nil {
+			// If detected error, check that error has nature of a normal behavior in the system
+			// and will not affect the further behavior
+			if err != nil && !(errors.Is(err, io.EOF) || strings.Contains(err.Error(), closedNetworkConnection)) {
 				errs = append(errs, err)
 			}
 		}
@@ -156,8 +172,10 @@ type leastConnection struct {
 	fwd *Forwarder
 }
 
+// Next will make selection of the next route using the algorithm of least
+// utilization (from the standpoint of this system) of the host connectivity
 func (lc leastConnection) Next() *route {
-	minVal := uint32(32000)
+	minVal := uint32(math.MaxUint32)
 
 	// Lock and unlock just to get access to the latest routes slice
 	// this delivers support for hot-reload of the routes by pointer refresh
