+++ tlsutil/tlsgen.go
@@ -1,4 +1,4 @@
-package testing
+package tlsutil
 
 import (
 	"fmt"
